+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Version 3.9
 - Fix a bug showing empty inventory column when exporting a formula
 - Improve formula is made feature
+- Add fixed header in formula view
 
 ### Version 3.8
 - tipsy.js dependency removed